+++ app/src/at-huynguyen/java/vn/asiantech/internship/unittest/UsernameValidation.java
@@ -1,4 +1,4 @@
-package vn.asiantech.internship;
+package vn.asiantech.internship.unittest;
 
 import java.util.regex.Pattern;
